Started the Project
Created the project in earnest and migrated some of the work done on a prototype project for the site. Installed only the required packages to make sure everything is done. The Homepage has aa good start but more work needs to be done.
The core toolkit for building the app and styling is set up. The groundwork is laid out.