Added Core Dark Mode
Implemented the logic for the Dark Mode switch. As such, from now on this project will no longer be using module.css sheets in most places because tailwind's dark mode clashes with the unique identifier files of module.css. Styling for most components, aside the hero banner, have been migrated inline and the files themselves deleted.

Dark mode overall works now, but on a basic level. Styling still needs to be set up for some current components and other components going forward. Core logic however is in place, so the hard part of Dark Mode as a whole is finished.

Also global.css background styling doesn't work with Tailwind dark mode either. So I created a background container instead as a work around and that will serve as the new background for the entire site. It is now layered in app.js.