Redux+Socket+Cleanup
Set the stage for Redux and the Redux Toolkit for the project's backend going forward but running npm i on those packages. Likewise did the same with Socket.io, but also tweaked package.json to make sure that when socket.io is implemented, it doesn't need to be configured. Hopefully.

Added a few other template files for new features like chatbox and chatinput so that a chat system can be implemented. However a chat system is a nice to have feature, not a core. This is just to potentially add it in the future. Can double up for notifications as well when users join a campaign.