Character sheet and Typescript Bugfix
Got a basic set up for the character sheet in preparation of incorporating the big 3; Redux, Socket.IO and SQL. The sheet looks nice but is basic in terms of content. Animations are in place and it is collapsible based on subsections. Installed a library for that purpose.

Furthermore, finally fixed the Typescript issue with <Input /> component from Material Tailwind. It was a bug/feature in that you needed to add CrossOrigin="" into the property list of the component itself, and Typescript stops complaining. Probably not working as intended and other ways to solve this Typescript gymnastics bug but whatever. This'll do.