File restructure and MySQL Foundation
Renamed files to follow a convention of lower case with a - in the middle for spaces. Has to be this way for ESL or whatever the hell. Maybe typescript. Also installed a package to talk to MySQL DB as well as set up the connection file + the environment variable file which will not be committed. Next up need to set up MySQL and the Schema itself to start actual connections to the DB and eventual wiring up to Redux.