Very basic Socket.io integration
Moved the test bed code I had from that dummy project for socket.IO into this project. Transition was NOT smooth. Now I need to run 2 servers for socket.io as it turns out. Core server on localhost 3000 and socket on 4000. This means come deploy, I'll need 3 services; core app, MySQL, and socket IO.

DigitalOcean is a good alternative because they can host everything; Next, SQL, socket. Only downside is they cost money but they do give 200$ free starter credit, and lowest tier is 7$ a month. So initial hosting on DigitalOcean will be good, at least for a bit. An eventual alternative will be better but who knows for now, worry about deploying when the time comes but for now, DigitalOcean will be the go-to.