Button perms and Alert polish
Set up button permissions on the navbar and the homepage. Also refactored the homepage so that it lives in its own homepage.tsx file instead of on the index.tsx page.

Things like login, register, logout now only displays when proper credentials are on the site. Also now using getStaticProps to protect pages on server side, with useEffect protecting client side. So double the protection. This is the format that will be used going forward everywhere except the about page.

Refined the alert component, it now should re-render less when not needed.