Refined Login and Dashboard
As it turns out, my login logic was set up poorly. Client session did happen, as did data fetching, but the server didn't establish a session token and thus the server would have no idea what state the client was in, and there would be a disconnect. After much trial and error, that was fixed and a session token is now made properly in my [...nextauth].ts route. As such, the login.ts was ditched because it is redundant.

Afterwards, finished implementing Dashboard logic. Data is now passed from the session token (user ID and such) and a dashboard is now generated per user, with data fetched from MySQL and properly poulated. Login and Daashboard components are now mostly done, can move onto the core functionality of the site soon. But need to typescript a lot of stuff beforehand, since the focus was on getting login and session tokens to work.