Added Page Fade Transitions and CSS Polish
Added some fade transitions when changing to a different page, installed and implemented Framer Motion for that. Nice library that may require a bit of extra overhead that doesn't clash with one other one that I tried. Had issue with routing clash cus it also used that. But now the site is overall more sleek and professional-looking.

Furthermore polished up and changed around some colors for the site to slightly reflect the sheets on table look. Still needs more refinement but as is its not bad and an improvement over previous colors.