Basic Redux/Thunk/Socket
Redux Observable was a dead end as it turns out. Support stopped around 2021 and their docs list code that's somewhat updated and would require a bit of fiddling to set up. Instead I went with redux thunk, which comes in the ReduxToolkit already.

Set that up alongside a very basic Socket.Io communication of having thunk/redux and Socket send test messages to the sever and client off of the handlesubmit in my core profile component. No actual data is being sent, yet, but the initial piping seems to be layed out well enough to build on top of that.