Socket/Redux/MySQL full functionality
IT. IS. DONE. One of the hardest logical puzzles to set up in this project, the groundwork is layed out for. On my character field I fetch info from my database, it populates the field, and most important if I have 2 tabs open, socket reflects what gets entered in one to other.

While there is still much to do in terms of site set up and creation of sheets, this is a pretty big piece to solve. Testing and scaling up will still need to be done, but, hopefully that won't introduce any major problems. Will need to document all this far better as well, but for now, the way this site will operate is no longer just a theory. It works.