Started Create Sheet
Got a lot done in terms of having a good foundation going for the Create Sheet aspect of the site. While internal logic is still almost completely absent, in terms of appearance it looks and behaves pretty good. Smooth, responsive, and will have tooltips here n' there.

Only core-profile is done, but it will serve as a great foundation going forward once internal logic is in place. Just need to start considering involving actual sessions and user data soon to tie the sheet to the user, and then work it all towards a submit into the database.