Finished setting up Create Sheet foundation
Aside from handling the overall submission of the form, and tying it to an account, I set up the necessary redux store as well as the logic to process the fields and make it so data is not lost when the user moves back and forth between subsections. Good progress overall but one thing I will need to get to work on immediately:
Refactoring the View portion. Right now naming conventions are inconsistent and I need to refactor that portion so it lines up with how my create sheet is structured now. I'll be committing here and then refactoring because renaming my stuff tends to always break it in fun and exciting ways.