Add Created Background
Added the create background subsection for the character sheet, as well as fine tuned subsections that contain all optional fields and figured out a map as to what sections need to go in what order. Now it's just a matter of rudimentary implementation. At least until the item section and spells section, I'll need to think about those a bit.