Finished CombatStats and Exploration Skills create sections
Got 2 more sections done for the create sheet component. The last of the biggest sections are now finished, the rest should be smaller in terms of size, just need to figure out some additional logic and formatting on how to handle things like items, gear, and spell slots. Could keep it simpler but we'll see. Also I'm taking a detour into fancier forrmatting when it comes to ability scores and the like. I want fancy +/- in front of inputs purely for visual flare but that's proving harder than anticipated.