Ground work for Skills and Passives create subsection.
Setting up a commit here to layout the SkillsPassives subsection. While I follow the exact same process as all the other subsections, the SkillsPassives subsection will require a different approach since the user will be able to add a number of input fields, to a limit, to accommodate the skills (and passives) they will to add.

This commit will serve as a foundation to then modify the logic, and front end, to accomplish this. One that is figured out, I can use this approach for spells, gear and items too.