Finished (Almost) the Skills section
Rebranded the SkillsPassives into just skills since technically passive aren't really a thing in DND. I have the vast majority of the functionality and styling going, which can be replicated into the spells, items and backpack sections.

Only thing missing is additional isValid validation. Right now I'm still using the checkbox method, which is fine, but name fields of each skill are required. Either I'll have to remove validation for that, or just keep as is. For now I won't sweat it too much. I'll give it some thought but frankly, it'll  work fine as is too.