Finished the core character sheet set up
Finished the vast majority of the redux store, styling and custom hooks for the remainder of the create sheet subsections. The "Homebrew" subsection is missing but that's fine, it's the least important one and most optional one, so for now it's not necessary to proceed.

What needs polishing is styling in some places, like spacing of components. Fleshing out the "Submit sheet" subcomponent and switching to a textarea instead of a text field in places like spells, items, and equipment. That won't take too long, but the MySQL schema will need to be tweaked a bit to accommodate such. I'll tweak that when I start getting the submersion process going which is up next; setting up the API route logic, submission logic and finally tying it to both a logged in user AND the database so that things can neatly transition into view sheet things.