Set up all section validation and initial data extraction for create sheet.
The submission subsection in create sheet now disables and enables button based on the other subsections being valid or not. Styling could use some work but core functionality is done. Also set up a custom hook that now pulls all data from all subsections and stores it in separate variables.

Though it remains to be seen if such extraction is correct. That's where setting up an API route comes in, though I'll have to find a way to test saving the data before hooking everything up to an actual logged in user.