Further route set up and validation
Refined the foundation for submitting data via API. Furthermore, extracted the Yup Schemas into their own file so that now they can serve as added validation not on just the UI side, but before the data is passed to axios to process. Having the Yup Validation schemas in their own files allows them to be used across the entire app, and not just in the create subcomponent hooks.

Finally, the routes will be split into each subsection because each will first be validated with the Yup Schema before sequentially being added to the database itself. Sheet creation still isn't linked to a profile, yet, but the POST request is much more structured and API logic is a lot closer to being done. After coreProfile logic is finished, it's just a matter of repeating the process across all other subsections.