Create Sheet user and database prep.
The core structure of the API route is set up with rate limiting, headers, and Yup Validation. Console logs display the proper data when a section is fully complete, and spits out nothing when its not so error handling will need to be put in place. That can be done soon, for now it is time to start putting it all together:
Linking character sheet to a user.
Sending data to a database.