Minor Commit for NextAuth migration
As it turns out, getSession will no longer work for "middleware", which included APIs. Which means the reason my API can't get a session right now is because getSession no longer does it. I need to use a different method as per NextAuth's docs; getServerSession, which should in theory do the trick.

Minor problem is it involves additional set up, which means tinkering with nextAuth's settings, which could break stuff. So a commit will be made here just in case.