CoreProfile API Route Almost Done
At long last, almost everything in terms of core POST logic for my API route is in place. I had to do quite a bit of restructuring to line things up with the newer version of NextAuth, headache really, but it paid off in that the user session can now be gotten in the API via getServerSession hook.

There was still one problem I couldn't overcome, no matter what I did, could not pull userId from serverSession. It seems NextAuth in its galaxy brained brilliance decides to keep things on the server side lightweight whether you like it or not, with name, email and image only, and those are mainly for OAuth stuff. I could of course tweak that library's innards and force it to take ID as well, but I found an easier workaround.

I simply grab id from getSession the custom hook that calls this API route, and others like it soon, and pass it in via a header. Then assign it to a variable in the route itself and voila, I have my ID finally. Not the most "secure" way to get ID, but the next best thing since I'm still using a session validator in the API route thanks to getServerSide, while grabbing my ID just outside the API call. Again, may not be the most secure way of doing it, but with all the other API route checks I have in place, it should suffice. Now all that's left is to make sure things get written into MySQL as they should, and then I can start doing this with all other 10 API routes. Once that's done, the character creator process is complete.