Finished create sheet API routes
Finished the remainder of the create sheet API routes as well as restructuring some of the MySQL schema to make it line up better with the data that's being sent, as well as deleting some of the older tables I thought I'd use like "backpack". For now my item/equipment system is simpler but we'll see, maybe I'll expand on it more but for now I have no plans.

Overall the logic behind sending a character sheet is done, I just need to do some finetuning and refinement on the API + the UI end. Next will likely come the create campaign section which should thankfully be much, MUCH smaller. After that I can start work on socket.io channels and linking everything together when it comes to view sheet which I still need to work on, but hopefully with all the things in place with create sheet that will go faster since I can re-use a bunch of things and not have to re-test API and MySQL data structure and junk.