Campaign List
Got quite a bit of work done over a few days in terms of a campaign list. Initially I wanted to keep it pretty basic, but discovered that Material UI had some pretty fancy list presets on offer, in terms of ui anyway. Logic would need to be done by me. I quite liked the UI on offer, so decided to put in the extra effort to have a fancier list menu.

And I did. While the API route logic is missing still, the route is there just need to be tested and wired up, the UX side of things is finished now. The list has a search function, it has pagination pages, and the pagination even adjusts itself based on search parameters. Pretty nifty! Best of all, I can use this fancy list over in my char sheet list as well.

On top of that, added a confirm dialogue box. A user, once the logic is in place, will be able to delete their sheet by clicking the trash icon. this will cause a dialogue to pop up and flash a warning that his action cannot be undone and if they're sure. The pop up can be reused, but the list itself will likely need to be copy pasted based on columns and stuff. Probably, we'll see. I'll look into re-usability, might be possible to make it more modular. But for now, an unexpected but nice addition of a fancier list.