Finished campaign list
The campaign list logic is done, and moved into its own custom hook for cleaner organization much like the other components. As is, I think I'm not going to strive to make this list modular but I will try to take out some bits and pieces and make them reusable, like the footer perhaps, we'll see. For now, this is for the most part ready to be replicated into a character sheet list too.