View List Refactoring
Got more work done for the campaign view list in that now it is indeed entirely modular. So I won't need to duplicate anywhere near as much code as I initially thought, this should neatly fit into character list, hopefully. There will be more rows there but we'll see, I'll tweak styling if need be or if completely necessary, duplicate just the body because the view list is now split into those 3 pieces; header, body, footer. If its not as compatible with character view as I thought, only the body will need to be tweaked so I can either break it up further or duplicate just that.