Submit Sheet Store Reset and View Campaign Initial Setup
Big commit involving 2 major things done;
One is the create sheet submission routes now have their store data reset, meaning that if a user submits a character sheet and goes to make another one right after, the data held by redux the first time around isn't kept. The store is wiped clean.

Second is I set up a get API route to retrieve campaign info as well as created a form to display such. Then I made it so in the view list component, the one that handles lists for both sheets and campaigns, I made it flexible enough so that now when you click on the text displayed it acts as a link that takes the user to the view page. This is only done on the campaign end right now, not the sheet end but adding it to the sheet end should be simple. Once the GET API route is made at least, which will be a lot less simple.

Next I need actually polish the view campaign form and heavily consider implementation of the final major piece; socket.IO. Because campaign sheets will act as a hot or a room for character sheets to pool off of, so that has to be implemented soon if not immediately.