Readme, about, and fine tuning
Greatly expanded the GitHub readme section to be more organized and informative, as well as added a series of very compressed images that will hopefully will display on Github's about project page. Won't know until I upload.

Furthermore added several tweaks to the site itself and added the about page. It is basic, but serves its purpose for now. Doesn't need to be elaborate either. Might add an image or two later.