Campaign View Page
Got a lot of work done setting up a campaign view page (and rolling back on Socket.IO) implementation. There were a few logical snags when it came to setting up that page, and I am not going to be implementing socket.IO to either character or campaign view right away, though I will set up my code as such to accommodate it for later. For now only view and onBlur stuff will be implemented.

While the view page is close to completion, the API and fetch logic is a bit needlessly complicated, at least for now it seems. There is a mis-align between campaign_ID and the rest of the data. This needs to be fixed and if possible I will simplify the fetching considerably. But just in case, gonna commit here cus it does, for the most part, work.