Dashboard Counters, Socket.IO disable and Some Cleanup
Added the campaign and sheet counters to dash board by fetching data from the user API route, now it displays numbers instead of being empty like before.

Furthermore; disabled Socket.IO stuff for the time being since it won't be a feature going forward for a while and thus, having the stuff set up is pointless especially since I'll be deploying on Vercel and not DigitalOcean to start.

Finally did some clean up as well like editing About page, adding links to the navbar which go to campaign list, and minor styling touch ups. Though more are in order in terms of dashboard for view campagin sheet for instance.