View Core Profile and potential Create Refactor
Working on the final component of view character sheet, and noticed some things I can reduce in my custom hook potentially when it comes to create character sheet. On top of that, might be able to use my custom hook in booth; view sheet and create sheet but that will require a lot of dangerous refactoring so as such, I am gonna commit here just in case despite view core profile not being quite done, but close.