Core profile view/create refactor completion
This for the most part should be it; core profile view is in entirely working order (after much fixing of stupid errors on DB side) as well as refactor of the creature subcomponent counterpart to make things more organized and less bloated. Create will still need to be quickly tested but outside of that, this format is what will be used to alter all the other subsections, both view and create, going forward. After that, it's almost time to deploy the base website.