Skills View/Create subsections done
Took far longer than I thought it would, was more difficult too since I had to accommodate a number update edge cases I haven't initially considered. This was my simplest type of component, until it turned into my most complex logic wise thanks to all the arrays and dynamic lists. But it is done, all I need to do now is replicate this in 3 other subcomponents; spells, items and equipment, and then that is it. Things are just about ready for MVP deploy.

I said these were going to be my simpler subcomponents to implement too, boy was that wrong. Was a whole thing, but this is it. Final stretch now.