Major Milestone: All 11 Create/View subsections finished
At long last, it is finally done. All 11 Create and View subsections have been refactored and made to be inline with the 1 custom hook design. With that, marks the completion of the development cycle for the "Minimal Viable Product" stage of this fantastic webzone.

Next up I can focus more on doing some testing, some fine tuning in terms of styles. There is much that can also be done in terms of code compression and refactoring in that manner, but that can be done bit by bit later on after deployment.

For now I will focus on tightening up a few obvious things and then deploying a version 1 on Vercel.