Dashboard User Info Refactor
Refactored and repaired my User Info dashboard component that displays UserId, date joined, total campaigns/sheets, etc. It was broken after I got rid of the Socket.IO stuff, and it was also designed differently to fetched API data and the like.

Now it is standardized to follow the format I use in my sheet and campaign routes. And it also works. Whether or not this will fix the routing issue in the deployed Vercel version I shall see after this commit. But the npm build version and the npm dev version both route just fine. So this leads me to believe that this routing issue is something environment variable related or... something. Who knows right now. Locally it completely works, but Vercel is different.