Post 1.0 deploy minor change
By refactoring and fixing my user dashboard and how I fetch userId, it fixed the Vercel issues as well in 1 fell swoop, which means that was indeed the root of the issue. And with that, version 1.0 of this fantastic webzone is complete.

There are many UI issues still left over, like the fact this site does not like Firefox. At all. The UI stuff just breaks in many sections like character list, so that will need to be looked at. Loading states will need to be added. Some of the character sheets polished a tad more. But other than that, for a portfolio site this will do.

I'll be making more improvements going forward, naturally, and this commit is to make the completion of 1.0. I guess its more of a Beta but for job searching its enough for now. More to come soon but I will be shifting focus away from this for a bit.