Additional Browser Support - Deploy Version 0.1.1
The long awaited feature of supporting other browsers, mainly Firefox, is at long last here.

In reality my styling already mostly supported browsers already. It was just a mild issue with core table styling that I had to fix. While resizing of a few other components is a bit scuffed, like on dashboard, the rest should work fine now and actually display across Chrome AND Firefox browsers.

As such, I'm removing mention of only chrome support and generalizing future styling improvements (which are still needed) as just UI/UX improvements. This is now version 0.1.1 of the site and will be committed so that the Vercel version is updated with these changes.